+++ js_modules/dagster-ui/packages/ui-core/src/graph/KindTags.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import {OpTags} from './OpTags';
 import {DefinitionTag} from '../graphql/types';
-import {linkToAssetTableWithKindFilter} from '../search/useGlobalSearch';
+import {linkToAssetTableWithKindFilter} from '../search/links';
 import {StaticSetFilter} from '../ui/BaseFilters/useStaticSetFilter';
 
 export const LEGACY_COMPUTE_KIND_TAG = 'kind';
